Refactor collections and subcollections with custom hooks
