+++ frontend/src/components/home/Collections.jsx
@@ -1,62 +1,25 @@
-import React, { useEffect } from "react";
+import React, { useRef } from "react";
 import { useNavigate } from "react-router-dom";
 import { motion, useInView } from "framer-motion";
-import { toast } from "react-toastify";
-import {
-  useGetCollectionQuery,
-  useGetSubCollectionsQuery,
-} from "@/redux/api/productApi";
 import CollectionGrid from "@/pages/Collections/CollectionGrid";
 import { CategoryFallback } from "@/components/product/shared/FallbackStates";
 import { categoryAnimations } from "@/hooks/Animation/animationConfig";
+import { useCollections } from '@/hooks/Product/useCollections';
 
 const Collections = () => {
   const navigate = useNavigate();
-  const ref = React.useRef(null);
+  const ref = useRef(null);
   const isInView = useInView(ref, { once: true, amount: 0.2 });
 
-  const { data, isError, error } = useGetCollectionQuery();
-  const {
-    data: subCollectionsData,
-    isError: isSubError,
-    error: subError,
-  } = useGetSubCollectionsQuery();
+  // Custom hook to get all collections
+  const { collections, handleCollectionClick } = useCollections(false);
 
-  // Get non-featured collections with null check
-  const collections =
-    data?.collections?.filter((c) => !c.isFeatured)?.slice(0, 6) || [];
-
-  useEffect(() => {
-    if (isError) {
-      toast.error(error?.data?.message || "Failed to fetch collections");
-    }
-    if (isSubError) {
-      toast.error(subError?.data?.message || "Failed to fetch sub-collections");
-    }
-  }, [error, isError, subError, isSubError]);
-
-  const handleCollectionClick = (collection) => {
-    if (!collection?._id) {
-      toast.error("Invalid collection");
-      return;
-    }
-
-    const hasSubCollections = subCollectionsData?.subcollections?.some(
-      (sub) => sub?.collection?._id === collection._id
-    );
-
-    // Always navigate to collections route if there are subcollections
-    if (hasSubCollections) {
-      navigate(`/collections/${collection._id}`);
-    } else {
-      // Navigate to products route if no subcollections
-      navigate(`/products?product_collection=${collection._id}`);
-    }
-  };
+  // Limit to 6 collections for homepage display
+  const limitedCollections = collections.slice(0, 6);
 
   return (
     <section ref={ref} className="p-4">
-      {collections.length > 0 ? (
+      {limitedCollections.length > 0 ? (
         <>
           <motion.h2
             variants={categoryAnimations.titleVariants}
@@ -83,7 +46,7 @@ const Collections = () => {
           </motion.div>
 
           <CollectionGrid
-            collections={collections}
+            collections={limitedCollections}
             onCollectionClick={handleCollectionClick}
             isInView={isInView}
           />
